Revert "Home styling2"
