+++ src/components/HorizontalScroller.js
@@ -9,7 +9,7 @@ import Grid from '@material-ui/core/Grid';
 import { Link } from "react-router-dom";
 import './HorizontalScroller.css';
 
-const MenuItem = ({film}) => {
+const MenuItem = ({film,filmlistId}) => {
   return (<div>
     
        <Grid item>
@@ -43,7 +43,7 @@ const MenuItem = ({film}) => {
 export const Menu = (list) =>
   list.map(film => {
  
-    return <MenuItem film={film} key={film.id}  />;
+    return <MenuItem film={film.film} filmlistId={film.id} key={film.film.id}  />;
   });
   const Arrow = ({ text, className }) => {
     return (
@@ -63,7 +63,6 @@ export const Menu = (list) =>
                         data={Menu(this.props.list)}
                         arrowLeft={ArrowLeft}
                         arrowRight={ArrowRight}
-                        alignCenter={false}
                         />
             
 
@@ -75,10 +74,19 @@ export const Menu = (list) =>
 
 const styles = {
   root: {
-    width: 120,
+    width: 150,
   },
   media: {
     height: 112.5,
   },
-  link: { textDecoration: 'none' }
+  link: { textDecoration: 'none' },
+  header: { width: 1000, margin: '0 auto', display: 'flex', flex: 1, flexDirection: 'column', justifyContent: 'center', padding: 20 },
+  link: { textDecoration: 'none' },
+  film: { width: 150, marginBottom: 15, marginRight: 10 },
+  stream: { width: 400 },
+  input: { border: 'none', backgroundColor: '#ddd', marginBottom: 10, padding: 8, fontSize: 18 },
+  filmTitle: { fontSize: 20, fontWeight: 'bold', marginBottom: 0 },
+  streamText: { fontSize: 14, marginBottom: 0},
+  filmDescription: { marginBottom: 0 }
+
 }
