--- src/components/HorizontalScrollerCircular.css
@@ -1,36 +1,0 @@
-
-.menu-item {
-    padding: 0 40px;
-    margin: 5px 10px;
-    user-select: none;
-    cursor: pointer;
-    border: none;
-     width:120px;
-  }
-  .menu-item-wrapper {
-    margin:0 10px;
-    width:120px
-  }
-   
-  .scroll-menu-arrow {
-    cursor: pointer;
-  }
-  .horizontal-menu{
-    width:100%;
-  }
-  .profile-img{
-    border-radius: 50%;
-    width:100px;
-    height:100px;
-    object-fit:cover
-  }
-  .profile-wrapper{
-    width: 120px;
-    overflow: hidden;
-    white-space: nowrap;
-    text-align: center;
-  }
-.pTitle{
-  text-align: center;
-  color:white
-}
